Add Dockerfile to build janet
This single Dockerfile can be used to build two versions of the janet
image, a minimal one with just janet and a sdk version with a compiler
and git so that we can build janet packages.

Which version of the image to build is set by the target option to
docker build - either `core` for the light version or `dev` for the sdk
version.

A particular commit of janet can be built bu passing in the build-arg
COMMIT.